Flagmania issues

+++ src/views/ExternalGameJoin.tsx
@@ -149,7 +149,7 @@ export const ExternalGameJoin = (props: PropsType) => {
       try {
         const res = await httpGet("/lobbies/" + lobbyId, { headers });
         const lobby = res.data as Lobby;
-        if (lobby.privateLobbyKey !== privateLobbyKey) {
+        if (!lobby.isPublic && lobby.privateLobbyKey !== privateLobbyKey) {
           setPlayer(undefined);
           setLobby(undefined);
           sessionStorage.clear();
@@ -226,7 +226,7 @@ export const ExternalGameJoin = (props: PropsType) => {
     } catch (error: any) {
       notifications.show({
         title: "Error",
-        message: error.errorresponse.data.message,
+        message: error.response.data.message,
         color: "red",
       });
       console.error();
